FIX: CLI: fix an exception when parsing scale params (#6969)
This PR aims to solve an exception when using OrcaSlicer in CLI which
caused the --scale option to not work.
All credits to @lanewei120 from BambuStudio for pushing it on BS repo.

Related commit on BS:

https://github.com/bambulab/BambuStudio/commit/2d4655e780e85e84445bb0b1987445e9e8d66b95#diff-2fa194122e892282ec589e51d5f7da65da83ec0488ced4b63930345d1273a353R3797
Related issue on BS:
https://github.com/bambulab/BambuStudio/issues/4628
Related issue on Orca:
https://github.com/SoftFever/OrcaSlicer/issues/6454
fixes #6454

```
./orca-slicer.exe --scale 2 --slice 1 --allow-newer-file test_model.stl --debug 5
```
The command above works again and scale the model then slice the scaled
model. Before, the scale would not complete.